assembler finished, debug needed
no crash error, could output file, but I haven't prove that the file is correct.
crash while free the space

+++ Planguage/PCoreBase.h
@@ -21,7 +21,7 @@ typedef struct AInstanceOfProcessor
   void *PID;
   unsigned int memoryUsage;
   void* APICallAddr;
-  unsigned long APICallID;
+  unsigned int APICallID;
   int status;
   void *pc;
   void *code;
